Added class for using local controller for use within pravega client
